tests/baseline/protocol-1.2/inspector/console/console-on-paint-worklet-expected.txt

CONSOLE INFO: line 1: Information
CONSOLE MESSAGE: line 2: Log
CONSOLE WARNING: line 3: Warning
Tests console output from PaintWorklet.

Message count: 3
console-worklet-script.js:1 Information
console-worklet-script.js:2 Log
console-worklet-script.js:3 Warning
(anonymous) @ console-worklet-script.js:3
